Refactor the code in a way that makes sense to me
Also I've changed it so that all the tile data is stored regardless of what's visible in the browser. This allows us to generate all the game data immediately, and we only reveal that data while playing. The original method generated the actual data while playing because the data only existed in the element data attributes, it wasn't stored on the tiles themselves.